refactor: change type of precipitationProbabilities prop in HourlyForecast to List<Float>

+++ app/src/main/java/com/example/justweather/data/remote/weather/models/HourlyWeatherInfoResponse.kt
@@ -26,7 +26,7 @@ data class HourlyWeatherInfoResponse(@Json(name = "hourly") val hourlyForecast: 
     @JsonClass(generateAdapter = true)
     data class HourlyForecast(
         @Json(name = "time") val timestamps: List<String>,
-        @Json(name = "precipitation_probability") val precipitationProbabilities: List<Double> = emptyList(),
+        @Json(name = "precipitation_probability") val precipitationProbabilities: List<Float> = emptyList(),
         @Json(name = "weathercode") val weatherCodes: List<Int> = emptyList(),
         @Json(name = "temperature_2m") val temperatureForecasts: List<Float> = emptyList()
     )
